Update links in analysis_options.yaml (#145706)
According to the [readme](https://github.com/dart-lang/linter?tab=readme-ov-file#repository-has-moved) of the linter repository, the source of truth for the linter has moved to the SDK repository. This updates the link to `all.yaml` in our file.

While I am in here, I also updated two other links that were outdated.